More debugging to track query state. Change the RequestState/RequestValue protocol to handle the case where there is nothing to send. Must return true only if a message was sent so the query state machine can track properly.

+++ cpp/src/command_classes/CommandClasses.cpp
@@ -235,7 +235,7 @@ void CommandClasses::ParseCommandClassOption
 	bool const _include
 )
 {
-	uint8 pos = 0;
+	size_t pos = 0;
 	uint8 start = 0;
 	bool parsing = true;
 	while( parsing )
